MoonCast/requirements.txt
torch==2.3.1
torchaudio==2.3.1
sentencepiece==0.2.0
protobuf
numpy

librosa==0.9.1
pyyaml
transformers
safetensors
einops
scipy
timm==1.0.7
torchdyn
librosa
accelerate==0.26.0
ninja
cryptography